requirements.txt

torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
transformers==4.30.2
accelerate==0.20.3
sentencepiece==0.1.99
protobuf==3.20.3
salesforce-lavis==1.0.2
ftfy==6.1.1
regex==2023.6.3
openai>=1.0.0
numpy==1.24.3
pandas==2.0.3
pyarrow==12.0.1
Pillow==9.5.0
pycocotools==2.0.6
scikit-learn==1.3.0
scipy==1.10.1
nltk==3.8.1
rouge-score==0.1.2
matplotlib==3.7.2
seaborn==0.12.2
tqdm==4.65.0
jupyter==1.0.0
notebook==6.5.4
ipywidgets==8.0.7
python-dotenv==1.0.0
pyyaml==6.0
requests==2.31.0
einops==0.6.1
timm==0.9.2
